win install script runs in user scope

+++ scripts/pslib/malmo_lib.psm1
@@ -205,24 +205,32 @@ function Install-VCRedist
 function Add-MalmoXSDPathEnv
 {
     $malmopath = $args[0]
-    [Environment]::SetEnvironmentVariable("MALMO_XSD_PATH", $malmopath + "\Schemas", "Machine")
+    [Environment]::SetEnvironmentVariable("MALMO_XSD_PATH", $malmopath + "\Schemas", "User")
     [Environment]::SetEnvironmentVariable("MALMO_XSD_PATH", $malmopath + "\Schemas", "Process")
 }
 
 function Install-Mesa
 {
-    Download-File "http://download.qt.io/development_releases/prebuilt/llvmpipe/windows/opengl32sw-64.7z" ($env:HOMEPATH + "\temp\mesa.7z")
-    & 'C:\Program Files\7-Zip\7z.exe' x .\temp\mesa.7z -o'.\temp\mesa' | Out-Host
-    if (-Not $?)
+    if (-Not (Test-Path ($env:HOMEPATH + "\temp\mesa.7z")))
     {
-        Write-Host "FAILED TO INSTALL MESA"
-        exit 1
+        Display-Heading "Installing Mesa software renderer"
+        Download-File "http://download.qt.io/development_releases/prebuilt/llvmpipe/windows/opengl32sw-64.7z" ($env:HOMEPATH + "\temp\mesa.7z")
+        & 'C:\Program Files\7-Zip\7z.exe' x .\temp\mesa.7z -o'.\temp\mesa' | Out-Host
+        if (-Not $?)
+        {
+            Write-Host "FAILED TO INSTALL MESA"
+            exit 1
+        }
+        mv .\temp\mesa\opengl32sw.dll .\temp\mesa\opengl32.dll
+        cp .\temp\mesa\opengl32.dll "$env:JAVA_HOME\bin"
+        if (-Not $?)
+        {
+            Write-Host "SOFTWARE RENDERER NOT ADDED TO JAVA HOME"
+            exit 1
+        }
     }
-    mv .\temp\mesa\opengl32sw.dll .\temp\mesa\opengl32.dll
-    cp .\temp\mesa\opengl32.dll "$env:JAVA_HOME\bin"
-    if (-Not $?)
+    else
     {
-        Write-Host "SOFTWARE RENDERER NOT ADDED TO JAVA HOME"
-        exit 1
+        Write-Host "Mesa already installed."
     }
 }
